requirements.txt
streamlit==1.39.0
plotly==5.24.0
requests==2.32.3
anthropic==0.97.0
httpx==0.28.1
python-dotenv==1.0.1
notion-client==3.1.0
pillow==10.4.0
pandas==2.3.3
